Seal ArrowPrimitiveType
